server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "openai>=2.21.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyjwt[crypto]>=2.11.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg-binary>=3.3.3",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.0",
    "mypy>=1.20.1",
    "pytest>=8.3.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 119
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
strict = true
python_version = "3.13"

warn_unreachable = true
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",    
]

local_partial_types = true

plugins = ["pydantic.mypy"]
exclude = ["migrations/"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = [
    "asyncpg",
    "asyncpg.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "langchain.*",
    "langchain_openai.*",
    "langchain_core.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["freezegun.*"]
ignore_missing_imports = true
